requirements.txt
fastapi
uvicorn
passlib
pymongo
motor
python-dotenv
pydantic[email]
python-jose
requests
